==> 1.1 Extract Restaurants.ipynb <==
# Initialize

import pandas as pd
import json
from geopy.distance import distance

# Load file
<li> load grab.csv containing all information of restaurants on grab</li>

grab_df = pd.read_csv("dataset/grab.csv")

# Filter for restaurants only
restaurants_df = grab_df[grab_df['loc_type']=='FOOD']

# filter relevant columns
restaurants_df = restaurants_df[['name', 'address','lat','lon', 'opening_hours']]
restaurants_df.head()

# Process opening_hours
<li>We want to filter for restaurants that are open on both Friday and Saturday past 12am</li>
<li>We want to filer for restaurants in Singapore's downtown region only</li>

# Extract friday and saturday opening hours
restaurants_df['fri_opening_hrs'] = restaurants_df['opening_hours'].apply(lambda x: json.loads(x)["fri"])
restaurants_df['sat_opening_hrs'] = restaurants_df['opening_hours'].apply(lambda x: json.loads(x)["sat"])

# Remove opening_hours column
restaurants_df = restaurants_df.drop('opening_hours', axis=1)

# Helper function to filter rows where opening hours extend past 00:00
def extends_past_midnight(opening_hrs):
    # Check if restaurants is closed
    if opening_hrs == 'Closed':
        return False
    
    time_range = opening_hrs.split()
    # Split the time range into opening and closing times
    def past_midnight(range):
        times = range.split('-')
        opening_time = times[0]
        closing_time = times[1]

        # Convert opening and closing times to datetime objects
        opening_datetime = pd.to_datetime(opening_time, format='%H:%M')
        closing_datetime = pd.to_datetime(closing_time, format='%H:%M')

        # Check if closing time is earlier than opening time, indicating it extends past 00:00
        return (closing_datetime < opening_datetime) or opening_time =="00:00"
    
    return any(past_midnight(range) for range in time_range)

# Create a binary variable that determines if a restaurant is open past 00:00
restaurants_df['fri_past_00:00'] = restaurants_df['fri_opening_hrs'].apply(extends_past_midnight)
restaurants_df['sat_past_00:00'] = restaurants_df['sat_opening_hrs'].apply(extends_past_midnight)

# Column to indentify restaurants that open past 00:00 on both friday and saturday
restaurants_df['midnight_restaurant'] = restaurants_df.apply(lambda x: True if x['fri_past_00:00'] and x['sat_past_00:00'] else False, axis=1)

# View of all restuarants open past 00:00 on both friday and saturday
restaurants_df[restaurants_df['midnight_restaurant']]

# Filter necessary columns
restaurants_df = restaurants_df[['name','address','lat','lon','midnight_restaurant']]

# Save dataframe as CSV
restaurants_df.to_csv("dataset/midnight_restaurants.csv", index=False)
