==> 1.2. Process Datasets.Rmd <==
---
title: "restaurants"
output: html_document
---

# 1. Loading all necessary libraries 
```{r setup}
library(GISTools)
library(tmap)
library(maptools)
library(sp)
library(sf)
library(dplyr)
library(osmdata)
```

# 2. Helper functions
```{r helper functions}
# Create plot of basemap with specified regions, nightlife locations in sg and tmap mode
nightlife_basemap_plot <- function(base_plg, nightlife_sf, mode) {
  tmap_mode(mode)
  tmap_options(check.and.fix = TRUE)

  tm_shape(base_plg) + tm_borders(col = "black") +
    tm_shape(nightlife_sf) + tm_dots(size = 0.005, col = "red") + # Nightlife locations
    tm_scale_bar(width = 0.15) +
    tm_layout(
      legend.format = list(digits = 0),
      legend.position = c("left", "bottom"),
      legend.text.size = 0.25,
      legend.title.size = 0.5,
      title = "Nightlife location in Singapore",
      title.position = c("left", "bottom")
    )
}
```

# 3. Loading and processing of datasets
```{r data loading}
# 2.1 Basemap - Singapore Subzones (Polygons Vectors)
subzones <- st_read(dsn = "dataset/basemap", layer = "MP14_SUBZONE_NO_SEA_PL")
base_map <- st_read(dsn = "dataset/basemap", layer = "SGP_adm0")
base_map <- base_map %>% select(ISO)
base_map <- st_transform(base_map, crs = st_crs(subzones))


# 2.2 Nightlife (Bars + Clubs) Data (Points Vectors)
query <- opq(bbox = "singapore, singapore") %>%
  add_osm_feature(key = "amenity", value = c("bar", "biergarten", "pub", "nightclub"))
rawdata_nightlife <- osmdata_sf(query)
nightlife_points <- rawdata_nightlife$osm_points %>%
  drop_na(name) %>%
  dplyr::select(name, amenity, `addr:postcode`, `addr:street`)

# some nightlife locations are represented as polygons, we convert them to point features
nightlife_plg <- rawdata_nightlife$osm_polygons %>%
  drop_na(name) %>%
  dplyr::select(name, amenity, `addr:postcode`, `addr:street`) %>%
  st_centroid()
nightlife <- rbind(nightlife_points, nightlife_plg)

# Create a Nightlife Spatial Points DataFrame
nightlife_sf <- st_transform(nightlife, crs = st_crs(subzones)) # project to basemap's CRS
nightlife_sf <- st_intersection(nightlife_sf, base_map) # remove nightlife points that lie outside of basemap


# 2.3 Restaurants
restaurants <- read.csv("dataset/midnight_restaurants.csv")
restaurant_coordinates <- cbind(restaurants$lon, restaurants$lat)

# Convert restaurants dataframe to a SpatialPointsDataFrame
restaurants_sp <- SpatialPointsDataFrame(restaurant_coordinates, data = data.frame(restaurants), proj4string = CRS("+proj=longlat +datum=WGS84"))
restaurants_sf <- st_as_sf(restaurants_sp)
restaurants_sf <- st_transform(restaurants_sf, crs = st_crs(subzones)) # project to basemap's CRS
restaurants_sf <- st_intersection(restaurants_sf, base_map) # remove restaurant points that lie outside of basemap

# View all nightlife locations in Singapore
nightlife_basemap_plot(base_map, nightlife_sf, "view")
nightlife_basemap_plot(base_map, restaurants_sf, "view")
```

# 4. Filter regions of interest (central regions of Singapore)
We are choosing to narrow our scope to nightlife locations in the central regions of Singapore. After close inspection of the basemap polygons, we define these central regions as Downtown Core, Singapore River, Rochor, Outram and Lavender.
```{r , echo=TRUE}
# Filter polygons of interest
View(subzones)
central_plg <- subzones %>% filter(PLN_AREA_C %in% c("DT", "SR", "RC", "OT") | SUBZONE_N %in% c("LAVENDER"))

# Plot with filtered polygons
nightlife_basemap_plot(central_plg, nightlife_sf, "view")
```

# 5. Filter nightlife and restaurant locations that lie in filtered polygons
```{r , echo=TRUE}
nightlife_filtered_sf <- st_intersection(nightlife_sf, central_plg)
restaurant_filtered_sf <- st_intersection(restaurants_sf, central_plg)

nightlife_basemap_plot(base_map, nightlife_filtered_sf, "view")
nightlife_basemap_plot(base_map, restaurant_filtered_sf, "view")
```

# 6. Save filtered nightlife and restaurants
```{r}
# Take not to delete files before overwriting
st_write(nightlife_filtered_sf, "dataset/processed_datasets/nightlife_locations.geojson")
st_write(restaurant_filtered_sf, "dataset/processed_datasets/restaurants.geojson")
```
